Modify program to use the new config providers
At this point we haven't deleted the old interfaces, so everything still runs, including the tests.

However, the program is using the new IConfiguration interface, and the rules and analyzer are still using the old IRuleParameterConfig interface.

To defer having to refactor the analyzer, rules and rule tests, we can temporarily create an adapter class that translates between IConfiguration and IRuleParameterConfig.